+++ interpreter/namerd/src/test/scala/io/buoyant/namerd/iface/NamerdTest.scala
@@ -9,8 +9,8 @@ import org.scalatest.FunSuite
 class NamerdTest extends FunSuite {
   test("sanity") {
     // ensure it doesn't totally blowup
-    val _ = NamerdInterpreterConfig(Some(Path.read("/whats/in/a")), Some("name"), None)
-      .interpreter(Stack.Params.empty)
+    val _ = NamerdInterpreterConfig(Some(Path.read("/whats/in/a")), Some("name"), None, None)
+      .newInterpreter(Stack.Params.empty)
   }
 
   test("interpreter registration") {