+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Removed unused TLS options from the k8s storage plugin config
 * Add an `ip` option to admin configuration so that access to the admin server may be constrained
 * Add k8s external namer for routing to k8s ingress services
+* Support serving the namerd namer interface over TLS.
 
 ## 0.8.1